requirements.txt
# Dev environment
pip
autopep8

# App

asgiref==3.7.2
autopep8==2.3.1
Django==5.0
django-cors-headers==4.5.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==11.0.0
pycodestyle==2.12.1
PyJWT==2.9.0
sqlparse==0.4.4
typing_extensions==4.9.0
